Fixed type warnings

+++ Library/src/Manager/AssetManager.cpp
@@ -17,7 +17,7 @@ bool AssetManager::ReadFolder(MW::ComPtr<ID3D11Device> &device, std::string path
 	{
 		std::string filepath = entry.path().string();
 		filepath.at(filepath.find('\\')) = '/';
-		int folderIndex = filepath.find_last_of('/');
+		size_t folderIndex = filepath.find_last_of('/');
 		std::string filename = filepath.substr(folderIndex + 1, filepath.length() - folderIndex);
 		std::string extension = filename.substr(filename.find_last_of('.'), 4);
